requirements.txt
streamlit
pinecone-client
google-generativeai
python-dotenv
langchain-core
langchain
langchain-community
langchain-pinecone
sentence-transformers
python-magic
python-magic-bin
pypdf
python-docx